Move radarr base entity to separate module
